+++ scripts/test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-target=${*:-tests.simple tests.proto tests.errorformat tests.conditional_rdeps tests.incremental}
+target=${*:-tests.simple tests.proto tests.errorformat tests.conditional_rdeps tests.incremental tests.python_imports}
 timeout_sec=100
 
 # Run full suite in multiprocessing mode, then in single-process mode